added an updateService
