+++ src/main/java/com/java/automation/lab/fall/antonyuk/core22/service/concreteService/ClientService.java
@@ -1,4 +1,4 @@
-package com.java.automation.lab.fall.antonyuk.core22.service;
+package com.java.automation.lab.fall.antonyuk.core22.service.concreteService;
 
 import com.java.automation.lab.fall.antonyuk.core22.domain.person.Client;
 
@@ -10,6 +10,7 @@ public class ClientService {
     public static synchronized <T> void update(Client client, String field, T value)
             throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         Method method = client.getClass().getMethod("set" + field, new Class[] {value.getClass()});
+        method.setAccessible(true);
         method.invoke(client, value);
     }
 }
