+++ src/main/java/com/java/automation/lab/fall/antonyuk/core22/service/concreteService/ExcursionService.java
@@ -1,7 +1,7 @@
-package com.java.automation.lab.fall.antonyuk.core22.service;
+package com.java.automation.lab.fall.antonyuk.core22.service.concreteService;
 
 import com.java.automation.lab.fall.antonyuk.core22.domain.club.HorseClub;
-import com.java.automation.lab.fall.antonyuk.core22.domain.event.TrainingGroup;
+import com.java.automation.lab.fall.antonyuk.core22.domain.event.Excursion;
 import com.java.automation.lab.fall.antonyuk.core22.domain.person.Client;
 import com.java.automation.lab.fall.antonyuk.core22.domain.person.Employee;
 
@@ -10,10 +10,11 @@ import java.lang.reflect.Method;
 import java.util.Date;
 import java.util.Set;
 
-public class TrainingGroupService {
-    public static synchronized <T> void update(TrainingGroup trainingGroup, String field, T value)
+public class ExcursionService {
+
+    public static synchronized <T> void update(Excursion excursion, String field, T value)
             throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-        Method method = trainingGroup.getClass().getMethod("set" + field, new Class[] {value.getClass()});
-        method.invoke(trainingGroup, value);
+        Method method = excursion.getClass().getMethod("set" + field, new Class[]{value.getClass()});
+        method.invoke(excursion, value);
     }
 }
